requirements.txt
pandas>=2.0,<3.0
openpyxl>=3.1
PyYAML>=6.0
